+++ src/js/blocks/course-continue-button/index.js
@@ -4,11 +4,15 @@
  * @since 1.0.0
  * @since 1.5.0 Add supported post type settings.
  * @since 1.8.0 Use imports in favor of "wp." variables.
+ * @version 2.5.0
  */
 
 // WP Deps.
 import { __ } from '@wordpress/i18n';
 import { Button } from '@wordpress/components';
+
+// Internal dependencies.
+import icon from '../../icons/forward';
 
 /**
  * Block Name
@@ -27,7 +31,8 @@ export const postTypes = [ 'course' ];
 /**
  * Register: Course Continue Button Block
  *
- * @since   1.0.0
+ * @since 1.0.0
+ * @since 2.5.0 Update icon color to `currentColor`.
  *
  * @param  {string}   name     Block name.
  * @param  {Object}   settings Block settings.
@@ -35,10 +40,7 @@ export const postTypes = [ 'course' ];
  */
 export const settings = {
 	title: __( 'Course Continue Button', 'lifterlms' ),
-	icon: {
-		foreground: '#2295ff',
-		src: 'migrate',
-	},
+	icon: icon,
 	category: 'llms-blocks', // common, formatting, layout widgets, embed. see https://wordpress.org/gutenberg/handbook/block-api/#category.
 	keywords: [ __( 'LifterLMS', 'lifterlms' ) ],