+++ src/js/blocks/form-fields/fields/user-address-street-primary.js
@@ -11,6 +11,7 @@ import { __ } from '@wordpress/i18n';
 // Internal Deps.
 import { settings as baseSettings, postTypes } from './text';
 import { getSettingsFromBase } from '../settings';
+import icon from '../../../icons/house';
 
 /**
  * Block Name
@@ -44,9 +45,7 @@ export const settings = getSettingsFromBase(
 			"A special field used to collect a user's street address.",
 			'lifterlms'
 		),
-		icon: {
-			src: 'admin-home',
-		},
+		icon: icon,
 		supports: {
 			multiple: false,
 			llms_field_inspector: {