+++ src/js/blocks/instructors/index.js
@@ -6,12 +6,14 @@
  * @since 1.0.0
  * @since 1.5.0 Add supported post type settings.
  * @since 1.11.0 Add membership to supported post types settings.
+ * @version 2.5.0
  */
 
 // WP Deps.
 import { __ } from '@wordpress/i18n';
 
-// Internal Deps.
+// Internal dependencies.
+import icon from '../../icons/person-chalkboard';
 import edit from './edit';
 import './editor.scss';
 
@@ -33,15 +35,13 @@ export const postTypes = [ 'course', 'llms_membership' ];
  * Register Block.
  *
  * @since 1.0.0
+ * @since 2.5.0 Update icon color to `currentColor`.
  *
  * @type {Object}
  */
 export const settings = {
 	title: __( 'Instructors', 'lifterlms' ),
-	icon: {
-		foreground: '#2295ff',
-		src: 'groups',
-	},
+	icon: icon,
 	category: 'llms-blocks',
 	keywords: [
 		__( 'LifterLMS', 'lifterlms' ),